refactor: extract WebMCP tools and simplify article rendering
