+++ package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@next/third-parties": "^16.2.9",
-    "@vercel/speed-insights": "^2.0.0",
     "next": "^16.2.9",
     "react": "^19.2.7",
     "react-dom": "^19.2.7"
